+++ src/main/java/com/motaharinia/msutility/custom/customvalidation/decimalcount/DecimalCountValidator.java
@@ -15,6 +15,8 @@ public class DecimalCountValidator implements ConstraintValidator<DecimalCount, 
     private static final String MESSAGE_EXACT = "CUSTOM_VALIDATION.DECIMAL_COUNT_EXACT";
     private static final String MESSAGE_MIN = "CUSTOM_VALIDATION.DECIMAL_COUNT_MIN";
     private static final String MESSAGE_MAX = "CUSTOM_VALIDATION.DECIMAL_COUNT_MAX";
+    private static final String UTILITY_EXCEPTION_MIN_OR_MAX_IS_NEGATIVE = "UTILITY_EXCEPTION.MIN_OR_MAX_IS_NEGATIVE";
+    private static final String UTILITY_EXCEPTION_MIN_IS_GREATER_THAN_MAX = "UTILITY_EXCEPTION.MIN_IS_GREATER_THAN_MAX";
 
     private String message;
     private Integer min;
@@ -42,21 +44,21 @@ public class DecimalCountValidator implements ConstraintValidator<DecimalCount, 
         if (exact > 0) {
             if (!exact.equals(decimalPart.length())) {
                 result = false;
-                message = MESSAGE_EXACT + "::" + exact;
+                setMessage(MESSAGE_EXACT + "::" + exact);
             }
         } else {
-            if (min <= 0 && max <= 0) {
+            if (min < 0 || max < 0) {
                 result = false;
-                message += "[min<=0 || max<=0]";
-            } else if (min > 0 && max > 0 && min > max) {
+                setMessage(UTILITY_EXCEPTION_MIN_OR_MAX_IS_NEGATIVE + "::min=" + min + "max=" + max);
+            } else if (min > max) {
                 result = false;
-                message += "[min>max]";
+                setMessage(UTILITY_EXCEPTION_MIN_IS_GREATER_THAN_MAX + "::min=" + min + "max=" + max);
             } else if (min > 0 && decimalPart.length() < min) {
                 result = false;
-                message = MESSAGE_MIN + "::" + min;
+                setMessage(MESSAGE_MIN + "::" + min);
             } else if (max > 0 && decimalPart.length() > max) {
                 result = false;
-                message = MESSAGE_MAX + "::" + max;
+                setMessage(MESSAGE_MAX + "::" + max);
             }
         }
         cvc.disableDefaultConstraintViolation();
@@ -64,5 +66,13 @@ public class DecimalCountValidator implements ConstraintValidator<DecimalCount, 
         return result;
     }
 
-
+    /**
+     * تنظیم پیام خطای پیش فرض در صورتی که توسعه دهنده پیام خاصی در انوتیشن ست نکرده باشد
+     * @param conditionalMessage خطای پیش فرض
+     */
+    private void setMessage(String conditionalMessage) {
+        if (ObjectUtils.isEmpty(message)) {
+            message = conditionalMessage;
+        }
+    }
 }