+++ src/main/java/com/motaharinia/msutility/custom/customvalidation/doublerange/DoubleRangeValidator.java
@@ -14,6 +14,8 @@ public class DoubleRangeValidator implements ConstraintValidator<DoubleRange, Do
 
     private static final String MESSAGE_MIN = "CUSTOM_VALIDATION.DOUBLE_RANGE_MIN";
     private static final String MESSAGE_MAX = "CUSTOM_VALIDATION.DOUBLE_RANGE_MAX";
+    private static final String UTILITY_EXCEPTION_MIN_OR_MAX_IS_NEGATIVE = "UTILITY_EXCEPTION.MIN_OR_MAX_IS_NEGATIVE";
+    private static final String UTILITY_EXCEPTION_MIN_IS_GREATER_THAN_MAX = "UTILITY_EXCEPTION.MIN_IS_GREATER_THAN_MAX";
 
     private String message;
     private Double min;
@@ -32,16 +34,28 @@ public class DoubleRangeValidator implements ConstraintValidator<DoubleRange, Do
             return true;
         }
         boolean result = true;
-        if (number < min) {
+        if (min > max) {
             result = false;
-            message = MESSAGE_MIN + "::" + min;
+            setMessage(UTILITY_EXCEPTION_MIN_IS_GREATER_THAN_MAX + "::min=" + min + "max=" + max);
+        } else if (number < min) {
+            result = false;
+            setMessage(MESSAGE_MIN + "::" + min);
         } else if (number > max) {
             result = false;
-            message = MESSAGE_MAX + "::" + max;
+            setMessage(MESSAGE_MAX + "::" + max);
         }
         cvc.disableDefaultConstraintViolation();
         cvc.buildConstraintViolationWithTemplate(message).addConstraintViolation();
         return result;
     }
 
+    /**
+     * تنظیم پیام خطای پیش فرض در صورتی که توسعه دهنده پیام خاصی در انوتیشن ست نکرده باشد
+     * @param conditionalMessage خطای پیش فرض
+     */
+    private void setMessage(String conditionalMessage) {
+        if (ObjectUtils.isEmpty(message)) {
+            message = conditionalMessage;
+        }
+    }
 }